extracting enter_wish to blueprint
